RunningTrials and LeagueOfPatience code completed
